# domain-customer-people-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/customers:
    put:
      tags:
      - Customers
      summary: Update customer
      description: Updates an existing customer
      operationId: updateCustomer
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - Customers
      summary: Register a customer
      description: Registers a customer and returns the resulting identifiers
      operationId: registerCustomer
      parameters:
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCustomerCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/phones/{phoneId}:
    put:
      tags:
      - Customers
      summary: Update customer email
      description: Modify phone fields keeping version history.
      operationId: updateCustomerPhone
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: phoneId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhoneCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Customers
      summary: Remove customer phone
      description: Remove phone from profile.
      operationId: removeCustomerPhone
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: phoneId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/emails/{emailId}:
    put:
      tags:
      - Customers
      summary: Update customer email
      description: Modify email fields keeping version history.
      operationId: updateCustomerEmail
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: emailId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Customers
      summary: Remove customer email
      description: Remove email from profile.
      operationId: removeCustomerEmail
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: emailId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/consents/{consentId}:
    put:
      tags:
      - Customers
      summary: Update customer consent
      description: "Upserts the consent record identified by (partyId, consentId).\
        \ Sets the granted flag and an optional applicationId soft link to the originating\
        \ loan or onboarding application."
      operationId: updateCustomerConsent
      parameters:
      - name: partyId
        in: path
        description: The owning party identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: consentId
        in: path
        description: The consent record identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsentCommand"
        required: true
      responses:
        "200":
          description: Consent updated
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid consent payload
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Party or consent record not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/addresses/{addressId}:
    put:
      tags:
      - Customers
      summary: Update customer address
      description: Modify address fields keeping version history.
      operationId: updateCustomerAddress
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: addressId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAddressCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - Customers
      summary: Remove customer address
      description: Remove or end-date an address; preserve audit.
      operationId: removeCustomerAddress
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: addressId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/customers/{partyId}/unlock:
    post:
      tags:
      - Customers
      summary: Unlock customer profile
      description: Unlock profile for changes.
      operationId: unlockCustomerProfile
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/reactivate:
    post:
      tags:
      - Customers
      summary: Reactivate customer
      description: Reactivate a dormant profile.
      operationId: reactivateCustomer
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/preferred-channel:
    post:
      tags:
      - Customers
      summary: Set preferred channel
      description: Set preferred contact channel (email/phone); one per type.
      operationId: setPreferredChannel
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferredChannelCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/phones:
    post:
      tags:
      - Customers
      summary: Add customer phone
      description: Add phone in E.164 with label (mobile/landline).
      operationId: addCustomerPhone
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPhoneCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/lock:
    post:
      tags:
      - Customers
      summary: Lock customer profile
      description: Lock profile for audit/investigation; block writes.
      operationId: lockCustomerProfile
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/id-documents:
    post:
      tags:
      - Customers
      summary: Add tax id
      description: Add secondary tax ID; validate format/uniqueness
      operationId: addTaxId
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterIdentityDocumentCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/emails:
    post:
      tags:
      - Customers
      summary: Add customer email
      description: Add email; validate format and uniqueness within customer.
      operationId: addCustomerEmail
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterEmailCommand"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/dormant:
    post:
      tags:
      - Customers
      summary: Mark customer dormant
      description: Flag profile as dormant due to inactivity.
      operationId: markCustomerDormant
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/deceased:
    post:
      tags:
      - Customers
      summary: Mark customer deceased
      description: Mark as deceased and block dependent operations.
      operationId: markCustomerDeceased
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/v1/customers/{partyId}/confirm-closure:
    post:
      tags:
      - Customers
      summary: Confirm customer closure
      description: Confirm closure after checks pass.
      operationId: confirmCustomerClosure
      parameters:
      - name: partyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    UpdateConsentCommand:
      type: object
      properties:
        partyId:
          type: string
          format: uuid
        consentId:
          type: string
          format: uuid
        granted:
          type: boolean
        applicationId:
          type: string
          format: uuid
...
    RegisterCustomerCommand:
      type: object
      properties:
        party:
          $ref: "#/components/schemas/RegisterPartyCommand"
        naturalPerson:
          $ref: "#/components/schemas/RegisterNaturalPersonCommand"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/RegisterPartyStatusEntryCommand"
        pep:
          $ref: "#/components/schemas/RegisterPepCommand"
        identityDocuments:
          type: array
          items:
            $ref: "#/components/schemas/RegisterIdentityDocumentCommand"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/RegisterAddressCommand"
        emails:
          type: array
          items:
            $ref: "#/components/schemas/RegisterEmailCommand"
        phones:
          type: array
          items:
            $ref: "#/components/schemas/RegisterPhoneCommand"
        economicActivities:
          type: array
          items:
            $ref: "#/components/schemas/RegisterEconomicActivityLinkCommand"
        consents:
          type: array
          items:
            $ref: "#/components/schemas/RegisterConsentCommand"
        providers:
          type: array
          items:
            $ref: "#/components/schemas/RegisterPartyProviderCommand"
        relationships:
          type: array
          items:
            $ref: "#/components/schemas/RegisterPartyRelationshipCommand"
        groupMemberships:
          type: array
          items:
            $ref: "#/components/schemas/RegisterPartyGroupMembershipCommand"
...